requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
requests==2.31.0
python-telegram-bot==20.4
prometheus-client==0.17.1
pydantic==2.0.3
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.5
